The gimp +++

+++ gen-libs/libunistring/01-libunistring-0.9.10.sh
@@ -2,7 +2,7 @@
 ${log} `basename "$0"` " started" blfs_all &&
 
 ${log} `basename "$0"` " download" blfs_all &&
-# if test -d /sources/libunistring-0.9.10
+if test -d /sources/libunistring-0.9.10
  then
   rm -rf /sources/libunistring-0.9.10
 fi
@@ -28,8 +28,7 @@ make &&
 ${log} `basename "$0"` " built" blfs_all &&
 
 make check &&
-${log} `basename "$0"` " unexpected check succeed" blfs_all
-${log} `basename "$0"` " expected check fail?" blfs_all &&
+${log} `basename "$0"` " check succeed" blfs_all &&
 
 make install &&
 ${log} `basename "$0"` " installed" blfs_all &&
